# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Neighbors",
        "version": "release 1.0"
    },
    "paths": {
        "/auth/login": {
            "post": {
                "tags": [
                    "AUTH"
                ],
                "summary": "Login que responde un Token",
                "description": "Login usando usuario y pass",
                "operationId": "authLogin",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "admin@neighbors.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "type": "string",
                                            "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJodHRwOlwvXC9sb2NhbGhvc3Q6NjA5MFwvYXV0aFwvbG9naW4iLCJpYXQiOjE2MTA5ODc0MDIsImV4cCI6MTYxMDk5MTAwMiwibmJmIjoxNjEwOTg3NDAyLCJqdGkiOiJlUXlTTzN2ZnVmY0UzS0F6Iiwic3ViIjoxLCJwcnYiOiIyM2JkNWM4OTQ5ZjYwMGFkYjM5ZTcwMWM0MDA4NzJkYjdhNTk3NmY3In0.HE5tmEaJ57_Z-PAgq9kTTQIvT_w7ycd-Hrtmez9YI2g"
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "example": "bearer"
                                        },
                                        "ok": {
                                            "type": "string",
                                            "example": "true"
                                        },
                                        "expires_in": {
                                            "type": "int",
                                            "example": 3600
                                        },
                                        "(obj)usuario": {
                                            "properties": {
                                                "habilitado": {
                                                    "type": "string",
                                                    "format": "boolean",
                                                    "example": "1"
                                                },
                                                "nombre": {
                                                    "type": "string",
                                                    "format": "text",
                                                    "example": "adminNombre"
                                                },
                                                "apellido": {
                                                    "type": "string",
                                                    "format": "text",
                                                    "example": "adminApellido"
                                                },
                                                "dni": {
                                                    "type": "int",
                                                    "format": "number",
                                                    "example": "87654321"
                                                },
                                                "telefono": {
                                                    "type": "int",
                                                    "format": "number",
                                                    "example": "12345678"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "format": "text",
                                                    "example": "admin@neighbors.com"
                                                },
                                                "img": {
                                                    "type": "string",
                                                    "format": "text",
                                                    "example": "mi_foto.jpg"
                                                },
                                                "uid": {
                                                    "type": "string",
                                                    "format": "text",
                                                    "example": "HQ21SD1R1F8S"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2021-01-18T15:31:11.000000Z"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2021-01-18T15:31:11.000000Z"
                                                },
                                                "id": {
                                                    "type": "int",
                                                    "format": "number",
                                                    "example": 2
                                                },
                                                "role": {
                                                    "properties": {
                                                        "0": {
                                                            "type": "string",
                                                            "format": "boolean",
                                                            "example": "ADMIN_ROL"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "(array)permisos": {
                                                    "properties": {
                                                        "(array)VECINO_ROL": {
                                                            "properties": {
                                                                "dashboard": {
                                                                    "properties": {
                                                                        "0": {
                                                                            "type": "string",
                                                                            "format": "text",
                                                                            "example": "read"
                                                                        }
                                                                    },
                                                                    "type": "object"
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "(array)barrios": {
                                                    "properties": {
                                                        "id_barrio": {
                                                            "type": "int",
                                                            "format": "number",
                                                            "example": "1"
                                                        },
                                                        "nombre_barrio": {
                                                            "type": "string",
                                                            "format": "text",
                                                            "example": "Jake O'Keefe"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "(array)viviendas": {
                                                    "properties": {
                                                        "id_lote": {
                                                            "type": "string",
                                                            "format": "text",
                                                            "example": "1"
                                                        },
                                                        "nombre_lote": {
                                                            "type": "string",
                                                            "format": "text",
                                                            "example": "miloteEjemplo1"
                                                        },
                                                        "nombre_barrio": {
                                                            "type": "string",
                                                            "format": "text",
                                                            "example": "Jake O'Keefe"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "email": {
                                            "type": "string",
                                            "example": "The email field is required."
                                        },
                                        "password": {
                                            "type": "string",
                                            "example": "The password field is required."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/auth/logout": {
            "post": {
                "tags": [
                    "AUTH"
                ],
                "summary": "Logout",
                "description": "Logout ( Primero autorizar los headers con el token JWT provisto en el login)",
                "operationId": "authLogout",
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User successfully signed out"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Error: Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Header autorization toker JWT",
                "name": "Token Bearer",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}